tracer_advection: review follow-ups of the WENO scheme package
- butterfly slot mask: compute_butterfly_slot_mask returns bool, the state and the hybrid
  selection stencil take it as an int32 0/1 field (was a float field);
- CellLocalPositiveDefinite checks at construction that the grid follows ICON's edge
  orientation convention, (geofac_div > 0) == (E2C[C2E][..., 0] == cell), which
  Jocksch's inline limiter assumes and the oriented port relies on
  (check_cell_edge_orientation_convention, unit-tested on the torus patch);
- AdvectionWenoHybridState documents that both states come from one least-squares
  setup and checks that their moment/geometry fields are the same objects;
- Fortran line references brought to branch transport_ajocksch_capture
  (mo_intp_coeffs_lsq_bln.f90: weight sets 2609-2622, slot spread 2631-2669, type-VI
  assembly 2670-2680, and the +5/+19/+23 shifts of the capture hooks elsewhere;
  mo_advection_hflux.f90: 103 routine 2532-3119, its limiter 3022-3042, the hybrid's
  unit weights 3684, residual 3547-3574); the paper is cited as Jocksch et al., PPAM 2026.

+++ model/atmosphere/tracer_advection/tests/tracer_advection/integration_tests/test_miura_weno_hybrid_pipeline.py
@@ -541,7 +541,7 @@ def _run_gt4py_hybrid_pipeline(
             for u in range(5)
         },
         lsq_butterfly_active=butterfly_field(
-            patch_coefficients["butterfly_active"].astype(np.float32)
+            patch_coefficients["butterfly_active"].astype(np.int32)
         ),
         use_weno=use_weno,
         selection_threshold=float(np.float32(threshold)),
